removed interger length from schema

+++ data/schema.sql
@@ -4,13 +4,13 @@ USE employee_tracker_db;
 
 
 CREATE TABLE department(
-  id INTEGER(11) AUTO_INCREMENT NOT NULL,
+  id INTEGER AUTO_INCREMENT NOT NULL,
   name VARCHAR(100),
   PRIMARY KEY (id)
 );
 
 CREATE TABLE role(
-  id INTEGER(11) AUTO_INCREMENT NOT NULL,
+  id INTEGER AUTO_INCREMENT NOT NULL,
   title VARCHAR(30),
   salary DECIMAL,
   department_id INTEGER,
@@ -19,7 +19,7 @@ CREATE TABLE role(
 );
 
 CREATE TABLE employee(
-  id INTEGER(11) AUTO_INCREMENT NOT NULL,
+  id INTEGER AUTO_INCREMENT NOT NULL,
   first_name VARCHAR(30),
   last_name VARCHAR(30),
   role_id INTEGER,
